spectr_utmc/controller_snapshot/home/voicelink/.releases/1.4.20/web_admin/db_repository/versions/001_migration.py
from sqlalchemy import *
from migrate import *


from migrate.changeset import schema
pre_meta = MetaData()
post_meta = MetaData()
camera = Table('camera', post_meta,
    Column('id', Integer, primary_key=True, nullable=False),
    Column('ip', String(length=16)),
    Column('alias', Integer),
    Column('number', Integer),
)


def upgrade(migrate_engine):
    # Upgrade operations go here. Don't create your own engine; bind
    # migrate_engine to your metadata
    pre_meta.bind = migrate_engine
    post_meta.bind = migrate_engine
    post_meta.tables['camera'].create()


def downgrade(migrate_engine):
    # Operations to reverse the above upgrade go here.
    pre_meta.bind = migrate_engine
    post_meta.bind = migrate_engine
    post_meta.tables['camera'].drop()
